backend/Goh/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf_spectacular==0.27.1
